requirements.txt
pyTelegramBotAPI==3.7.4
python-binance==0.7.5
python-dotenv==0.15.0
APScheduler==3.6.3
pandas
numpy